Move all listeners to its own subpackage
Also move build permission message and death message into the main properties file
